Move declaration IScriptManager to ScriptManager.cs
