+++ GitUI/GitUI.csproj
@@ -456,7 +456,6 @@
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="Hotkey\KeysExtensions.cs" />
-    <Compile Include="Script\IScriptManager.cs" />
     <Compile Include="Script\IScriptOptionsParser.cs" />
     <Compile Include="Script\IScriptRunner.cs" />
     <Compile Include="Script\PowerShellHelper.cs" />
