--- GitUI/Script/IScriptManager.cs
@@ -1,17 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace GitUI.Script
-{
-    internal interface IScriptManager
-    {
-        BindingList<ScriptInfo> GetScripts();
-
-        ScriptInfo GetScript(string key);
-
-        void RunEventScripts(GitModuleForm form, ScriptEvent scriptEvent);
-
-        string SerializeIntoXml();
-
-        int NextHotKeyCommandIdentifier();
-    }
-}
